update lobby and player screens
