--- front-end/src/components/SpectatorView.tsx
@@ -1,1 +1,0 @@
-//Nothing here yet, thinking of having spectator spectate in lobby screen
