+++ front-end/src/pages/Lobby/index.tsx
@@ -6,13 +6,13 @@ import WebSocketService from "../../services/WebSocketService.ts";
 import type { GameMessage } from "../../services/WebSocketService.ts";
 import { useNavigate } from 'react-router-dom';
 
+//TODO: Add styling to page
 const Index: React.FC = () => {
   //
   const { lobby, user } = useGame();
   const [lobbyDetails, setLobbyDetails] = useState<Lobby | null>(null);
   const [messages, setMessages] = useState<GameMessage[]>([]);
   const messagesEndRef = useRef<HTMLDivElement>(null);
-  const [game_status] = "status_here"; //placeholder for game_status later
   const navigate = useNavigate();
 
   const fetchLobbyDetails = async (lobbyCode: string) => {
@@ -37,14 +37,16 @@ const Index: React.FC = () => {
 
     if (!lobby?.code) return;
     fetchLobbyDetails(lobby.code);
-    //const interval = setInterval(() => fetchLobbyDetails(lobby.code!), 2000);
+    
     if(!WebSocketService.isConnected())
     {
       //Open new connection when not connected
       WebSocketService.connect(lobby?.code, user.teamId, user.id, handleGameMessage);
       console.log('Executed one to 3');
     }
-    
+    //We don't need to poll on certain intervals, the websocket will tell us when to
+    //- We just listen for a "join" message and poll
+    //const interval = setInterval(() => fetchLobbyDetails(lobby.code!), 2000);
     //return () => clearInterval(interval);
   }, [lobby?.code, user?.teamId, user?.id]);
 
@@ -54,8 +56,9 @@ const Index: React.FC = () => {
     //Handle different message types
     switch (msg.type) {
       case 'start_game':
-        //Redirect logic would go here
         //console.log("Start game received, would redirect now");
+        //Redirect to player scree page
+        //-They cannot go back to previous page
         navigate('/player', {replace: true})
         break;
       case 'join':
@@ -66,12 +69,13 @@ const Index: React.FC = () => {
         break;
       default:
         //For other message types, just add to messages
+        //-This is just incase, but the only messages to be recieved at this stage are for "join" and "start_game"
         setMessages(prev => [...prev, msg]);
         break;
     }
   };
 
-  // Format message for display based on type
+  //Format message for display based on type
   const formatMessage = (msg: GameMessage): string => {
     switch (msg.type) {
       case 'join':
@@ -83,42 +87,42 @@ const Index: React.FC = () => {
 
   if (!lobbyDetails) {
     return (
-      <div className="min-h-screen bg-gradient-to-br from-indigo-900 to-purple-800 flex items-center justify-center p-4">
-        <div className="bg-gray-800/80 backdrop-blur-sm rounded-xl p-8 text-center text-white shadow-2xl border border-purple-500/30">
-          <h2 className="text-xl font-bold">Loading Lobby Details...</h2>
-          <p className="mt-2 text-purple-200">Please wait while we connect you to the game</p>
+      <div>
+        <div>
+          <h2>Loading Lobby Details...</h2>
+          <p>Please wait while we connect you to the game</p>
         </div>
       </div>
     );
   }
 
-  // Use new backend structure for teams
+  //Use new backend structure for teams
   const teams =
     typeof lobbyDetails.teams === "object" && "teams" in lobbyDetails.teams
       ? Object.values((lobbyDetails.teams as { teams: Record<string, any> }).teams ?? {})
       : [];
 
-  return (
-    <div className="min-h-screen bg-gradient-to-br from-indigo-900 to-purple-800 p-4 md:p-6">
+ return (
+    <div className="min-h-screen p-4 md:p-6">
       {/* Header Section */}
-      <div className="bg-gray-800/80 backdrop-blur-sm rounded-xl p-6 mb-6 text-white shadow-2xl border border-purple-500/30">
+      <div className="rounded-xl p-6 mb-6 shadow-2xl">
         <div className="flex flex-col md:flex-row justify-between items-center">
           <div>
-            <h1 className="text-2xl md:text-3xl font-bold text-transparent bg-clip-text bg-gradient-to-r from-cyan-400 to-purple-400">
+            <h1 className="text-2xl md:text-3xl font-bold">
               Game Lobby
             </h1>
-            <p className="text-purple-200 mt-1">Waiting for players to join...</p>
+            <p className="mt-1">Waiting for players to join...</p>
           </div>
           <div className="mt-4 md:mt-0 text-center md:text-right">
-            <div className="text-sm text-purple-300">Lobby Code</div>
-            <div className="text-2xl font-mono font-bold bg-purple-900/50 px-4 py-2 rounded-lg mt-1 border border-purple-500/50">
+            <div className="text-sm">Lobby Code</div>
+            <div className="text-2xl font-mono font-bold px-4 py-2 rounded-lg mt-1">
               {lobby?.code}
             </div>
           </div>
         </div>
         
         <div className="mt-4 flex items-center">
-          <div className={`h-3 rounded-full ${game_status === "playing" ? 'bg-green-500' : 'bg-amber-500'} w-3 mr-2`}></div>
+          <div className={`h-3 rounded-full w-3 mr-2`}></div>
           <span className="text-sm">
             Status: {"game_status" in (lobbyDetails.teams as any)
               ? (lobbyDetails.teams as { game_status: string }).game_status
@@ -131,37 +135,37 @@ const Index: React.FC = () => {
         {/* Teams Section */}
         <div className="lg:col-span-2 grid grid-cols-1 md:grid-cols-2 gap-4">
           {teams.map((team: any, idx: number) => (
-            <div key={team.id} className="bg-gray-800/80 backdrop-blur-sm rounded-xl p-5 text-white shadow-2xl border border-purple-500/30">
+            <div key={team.id} className="rounded-xl p-5 shadow-2xl">
               <div className="flex justify-between items-start mb-4">
                 <h3 className="text-xl font-bold">
                   Team {team.color} {team.shape}
                 </h3>
-                <div className="bg-purple-700/50 px-3 py-1 rounded-full text-sm">
+                <div className="px-3 py-1 rounded-full text-sm">
                   {team.players?.length || 0}/{team.max_players}
                 </div>
               </div>
               
               <div className="mb-4 flex items-center">
-                <div className="text-amber-300 font-bold text-lg mr-2">Score: {team.score ?? 0}</div>
-                <div className="text-sm text-purple-300 ml-auto">
+                <div className="font-bold text-lg mr-2">Score: {team.score ?? 0}</div>
+                <div className="text-sm ml-auto">
                   Needs: {team.max_players - (team.players?.length || 0)}
                 </div>
               </div>
               
-              <div className="bg-gray-900/50 rounded-lg p-3 mb-3">
-                <h4 className="text-sm font-semibold text-purple-300 mb-2 border-b border-purple-900 pb-1">Players</h4>
+              <div className="rounded-lg p-3 mb-3">
+                <h4 className="text-sm font-semibold mb-2 pb-1">Players</h4>
                 <div className="space-y-2">
                   {(team.players ?? []).map((player: any) => (
-                    <div key={player.id} className="flex items-center bg-gray-800/40 p-2 rounded-md">
-                      <div className="w-2 h-2 bg-green-500 rounded-full mr-2"></div>
+                    <div key={player.id} className="flex items-center p-2 rounded-md">
+                      <div className="w-2 h-2 rounded-full mr-2"></div>
                       <div>
                         <div className="font-medium">{player.name}</div>
-                        <div className="text-xs text-gray-400">ID: {player.id}</div>
+                        <div className="text-xs">ID: {player.id}</div>
                       </div>
                     </div>
                   ))}
                   {(!team.players || team.players.length === 0) && (
-                    <div className="text-center text-gray-400 py-3">Waiting for players...</div>
+                    <div className="text-center py-3">Waiting for players...</div>
                   )}
                 </div>
               </div>
@@ -170,16 +174,16 @@ const Index: React.FC = () => {
         </div>
 
         {/* Messages Section */}
-        <div className="bg-gray-800/80 backdrop-blur-sm rounded-xl overflow-hidden shadow-2xl border border-purple-500/30">
-          <div className="bg-gradient-to-r from-purple-700 to-indigo-700 px-5 py-3">
-            <h3 className="text-lg font-bold text-white flex items-center">
+        <div className="rounded-xl overflow-hidden shadow-2xl">
+          <div className="px-5 py-3">
+            <h3 className="text-lg font-bold flex items-center">
               <span className="mr-2">💬</span> Lobby messages
             </h3>
           </div>
           
-          <div className="h-96 overflow-y-auto p-4 bg-gray-900/40">
+          <div className="h-96 overflow-y-auto p-4">
             {messages.length === 0 ? (
-              <div className="text-gray-400 text-center py-8 flex flex-col items-center">
+              <div className="text-center py-8 flex flex-col items-center">
                 <p>Welcome to the lobby!</p>
                 <p className="text-sm mt-1">Messages will appear here</p>
               </div>
@@ -187,9 +191,9 @@ const Index: React.FC = () => {
               <div className="space-y-3">
                 {messages.map((msg, index) => (
                   <div key={index} className="animate-fadeIn">
-                    <div className="text-sm bg-gray-800/60 p-3 rounded-lg border-l-4 border-purple-500">
-                      <div className="text-white">{formatMessage(msg)}</div>
-                      <div className="text-xs text-gray-400 mt-1">Just now</div>
+                    <div className="text-sm p-3 rounded-lg border-l-4">
+                      <div>{formatMessage(msg)}</div>
+                      <div className="text-xs mt-1">Just now</div>
                     </div>
                   </div>
                 ))}
@@ -198,8 +202,8 @@ const Index: React.FC = () => {
             )}
           </div>
           
-          <div className="border-t border-gray-700 p-3 bg-gray-800">
-            <div className="text-xs text-gray-400 text-center">
+          <div className="border-t p-3">
+            <div className="text-xs text-center">
              Laser shooter • {teams.reduce((acc, team) => acc + (team.players?.length || 0), 0)} players online
             </div>
           </div>
@@ -209,10 +213,10 @@ const Index: React.FC = () => {
       {/* Game Start Indicator */}
       {"game_status" in (lobbyDetails.teams as any) && 
         (lobbyDetails.teams as { game_status: string }).game_status === "starting" && (
-        <div className="fixed inset-0 bg-black/70 flex items-center justify-center z-50">
-          <div className="bg-gray-800 p-8 rounded-xl text-center animate-pulse border border-cyan-500/50">
-            <h2 className="text-2xl font-bold text-cyan-400 mb-2">Game Starting Soon!</h2>
-            <p className="text-gray-300">Get ready to play</p>
+        <div className="fixed inset-0 flex items-center justify-center z-50">
+          <div className="p-8 rounded-xl text-center animate-pulse">
+            <h2 className="text-2xl font-bold mb-2">Game Starting Soon!</h2>
+            <p>Get ready to play</p>
           </div>
         </div>
       )}
